Rename educational components
